fixing the services, fixing types and adding configuration

+++ NextGen-BM-BE/NextGen-BM-BE-Domain/Entities/BuildingAggregate/Building.cs
@@ -6,7 +6,7 @@ public class Building
     public int AddressId { get; set; }
     public string? Alias { get; set; }
     public int FloorNum { get; set; }
-    public double TotalBuildingSize { get; set; }
+    public decimal TotalBuildingSize { get; set; }
     public DateOnly DateBuilt { get; set; }
     public int NumOfElevators { get; set; }
     public required Address BuildingAddress { get; set; }
